# Conflicts:
#	src/main/java/com/northwind/controllers/HomeController.java
#	src/main/java/com/northwind/controllers/ProductController.java
#	src/main/java/com/northwind/repositories/ProductRepository.java

+++ src/main/java/com/northwind/entities/OrderDetail.java
@@ -8,7 +8,7 @@ public class OrderDetail {
 
     public int productID;
     public double unitPrice;
-    public short quantity;
+    public int quantity;
     public String discount;
 
     public int getProductID() {
@@ -27,11 +27,11 @@ public class OrderDetail {
         this.unitPrice = unitPrice;
     }
 
-    public short getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(short quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
